Share the trimmers' segments and timeline, and keep what made each tool's own
The first of the near-duplicate leaves: the audio trimmer was written from the
video cutter, and its segments.js and timeline.js were that tool's with the
nouns changed and three things genuinely different. Those three are now the
whole of what each tool keeps.

segments.js - the marks and the text file they are saved as - was the same
file apart from one number, the shortest segment worth keeping: a
millisecond of sound is forty-eight samples and a millisecond of video is no
frame at all. The shared file takes that as a parameter, and each tool's
segments.js is eight lines naming its number, with the reasoning beside it,
and passing it on. The audio trimmer's keeps formatDuration as well, which
only its rows show.

timeline.js - the bar with every segment drawn on it - differed in what lies
under the bands, where a mark lands, and how far the fine controls step.
The video cutter draws a tick at every keyframe and lands every mark on the
nearest frame; the audio trimmer draws the waveform and lets a mark sit on
any sample. So the shared bar takes the layer as an element, calls
decorate() on it when the source changes, and has a snap() a subclass may
override, and each tool's timeline.js is a subclass of thirty lines that
says exactly that. The band's tooltip is the page's to word: the audio
trimmer says it through its tl.band phrase as before, and the video cutter
still says it in English, now from main.js, where it can be seen to be the
one sentence on that page that is not yet a phrase.
